refractored cart related

+++ src/application/services/cart/cart.service.ts
@@ -1,24 +1,22 @@
 import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { plainToInstance } from 'class-transformer';
 
-import { CartItemDTO } from '../../dto/cart/cart-item.dto';
-import { CartItemResponseDTO } from '../../dto/cart/cart-item.response.dto';
+import { ItemDTO } from '../../dto/item/item.dto';
 import { CartResponseDTO } from '../../dto/cart/cart.response.dto';
-import { RemoveItemResponseDTO } from '../../dto/cart/remove-item.response.dto';
+import { CartItemResponseDTO } from 'src/application/dto/cart/cart-item.dto';
 
 import { ICartRepo } from '../../../domain/interfaces/repositories/cart.repository.interface';
 import { ICartService } from '../../../domain/interfaces/services/cart.service.interface';
-import { ProductService } from '../product/product.service';
-
+import { IProductService } from 'src/domain/interfaces/services/product.service.interface';
 
 @Injectable()
 export class CartService implements ICartService{
   constructor(
     private readonly cartRepo: ICartRepo,
-    private readonly productService: ProductService
+    private readonly productService: IProductService
   ) {}
   
-  async addItemToCart(userId: string, dto: CartItemDTO): Promise<CartItemResponseDTO> {
+  async addItem(userId: string, dto: ItemDTO): Promise<CartItemResponseDTO> {
     try {
       const productVariant = await this.productService.findProductVariant(dto);
       /* Simple request validation */
@@ -36,7 +34,7 @@ export class CartService implements ICartService{
 
       const cartItem = await this.cartRepo.upsertCart(cart.id, dto);
 
-      return new CartItemResponseDTO(cartItem);
+      return new CartItemResponseDTO(cartItem); 
 
     } catch (error) {
       console.error(`Error adding item to cart for user ${userId}:`, error);
@@ -49,14 +47,46 @@ export class CartService implements ICartService{
   };
 
   async getCart(userId): Promise<CartResponseDTO> {
-    const cart = await this.cartRepo.getCartDetails(userId);
-    return plainToInstance(CartResponseDTO, cart, {
+    try {
+      let cart = await this.cartRepo.findCartByUserId(userId);
+      if (!cart) {
+        cart = await this.cartRepo.createCart(userId);
+      }
+
+      cart = await this.cartRepo.getCartDetails(userId);
+
+      return plainToInstance(CartResponseDTO, cart, {
       excludeExtraneousValues: true,
-    });
+      });
+
+    } catch (error) {
+      console.error(`Error fetching cart for user`, error);
+
+      if (error instanceof NotFoundException || error instanceof BadRequestException) {
+        throw error; // Re-throw the specific error
+      }
+        throw new BadRequestException('An unexpected error occurred during fetching cart.');
+    }
   };
 
-  async removeItem(userId: string, cartItemId: string): Promise<RemoveItemResponseDTO> {
-    const cartItem = await this.cartRepo.removeCartItem(userId, cartItemId);
-    return new CartItemResponseDTO(cartItem);
+  async removeItem(userId: string, cartItemId: string): Promise<CartItemResponseDTO> {
+    try {
+      let cartItem = await this.cartRepo.findItem(cartItemId);
+      if (!cartItem) {
+        throw new BadRequestException(`There's no item id: ${cartItemId} in your cart`);
+      }
+      cartItem = await this.cartRepo.removeCartItem(userId, cartItemId);
+
+      return new CartItemResponseDTO(cartItem);
+
+    } catch (error) {
+      console.error(`Error removing item from cart for user`, error);
+
+      if (error instanceof NotFoundException || error instanceof BadRequestException) {
+        throw error; // Re-throw the specific error
+      }
+        throw new BadRequestException('An unexpected error occurred during removing item from cart.');
+    }
+    
   };
 }
